allow rate-limiting on data nodes (for shards.tolerant=true)
this is a quick-and-dirty hack that will work for our
usage, but this should be reconsidered and something
more general committed upstream.

Honestly it doesn't make sense to hardcode handling of
this at the level of RateLimitManager. Individual rate
limiters should be specified as plugins, with the
context necessary to make their own decisions.